# Fixed bug when parsing an unknown value in a proto2 enum extension.  …